refactor(handlers): extract diagnose.ts handler from intent.ts (Stage B)
handleSAPDiagnose + dump-section helpers. Also relocates isBtpSystem to
feature-cache.ts (cached-state helper shared by read+diagnose). 3672 tests,
A1 zero-diff.
